build: Bake git hash version into the service docker image
We want to be able to rollout images and rollout the bootstrap version we serve
when doing it.

To do so, we give the bootstrap hash as parameter to service as an argument in
the docker image (and not, say, in a kubernetes manifest). Bake the current
HEAD into that Dockerfile.service, generating it at build time.