sentry: Factor out a tiny helper for logging to sentry
This is a bit fiddly if we want the correct stack trace and a good message.

We also don't want to use the Exception sentry object which, with the golang
library, ends up with the name of error type as message prefix. The issue with
that is, when only using error.New(), all errors have the same prefix.