kubectl: Simplify parseOutputVersion
We were using struct to store client and server versions, but this was from the
time where we decoded some JSON output and used those structs to unmarshal
into.

Lili pointed out we can do things in a nicer way now that we don't parse JSON,
removing the usage of those structs entirely.

Still need to be a bit careful with the error management, in particular kubectl
version could return the client version but not the server version because it
couldn't contact the API servers.