bootstrap: Bump the DNS test timeout to 1m
We have seen in the wild a otherwise perfectly working k8s hit the 30s timeout
(it typically takes 3s in our dev cluster). When the check is skipped the rest
of the installation happens flowlessly.

Bump the timeout to 1m to leave more time for the test to run.