agent: Fix agent poll URL with CRI endpoints
When we have an agent with a CRI endpoint specified, the auto-update has to
cary that state over. We do that by appending a GET parameter to the URL
generating the agent manifest.

An earlier attempt, changing the default agent poll URL didn't work because we
override that URL in the manifest served by service, removing the bit of
conditional templating that was added to the default parameter value.

Fixes: #256