agent: Correctly encode the new GET parameters
I thought the library would do it for me, but no. The URL struct only has the
`WawQuery` field and String() compute the URL string with that field.

To add a GET parameter we need to encode the RawQuery manually.

This time, I unit-tested it, that's the price to pay when not getting right the
first time!