Simplify the flux flags parsing
We can save the clean-up phase, which reset things to `nil` if they
weren't passed as flags, by checking with the flag parser
directly. This makes it trickier to construct a config (since it needs
the flags parser state), but that's what `ParseFluxArgs` is for.

The code that uses ParseFluxArgs is sensitive to whether it gets a
`nil`, a value, or an error, so that behaviour must be preserved.