build: Add timeout to curl loops
If minikube were to return something rubbish, curl would still try to connect to
it - and it might wait for a reply for quite some time. It might be better to
back off and try again, which is why I've added a maximum 1 second request time.

Also removing the pipe all output to /dev/null - it won't normally print
anything (it's in quiet mode) but if we *were* to do something horribly wrong,
I would like to see it.