Add github action to mirror from dockerhub to ghcr
Scanning of dockerhub for new images is causing us to hit their rate
limit.

This github action uses google's 'Crane' tool[1] to copy the versions of
the images that we need from dockerhub to github container registry
(ghcr).

A third party action (imjasonh/setup-crane) is used to install crane
with the correct version.

[1] https://github.com/google/go-containerregistry/tree/a0c4bd256482b8522065d5f6cf966281ef270680/cmd/crane