ci: Upgrade kind version used in tests, stop testing k8s 1.12
The latest version can no longer be provisioned with the old kind
release, so we need a newer one.

The oldest version of k8s that we were testing, 1.12, started timing
out a few days ago and won't work again. I'm not sure why, we didn't
make any changes, so it must be something else that changed. However,
we don't really support 1.12 anyway, so we might as well drop it.